+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Hide plot_number in plots view. [#634](https://github.com/Rothamsted-Ecoinformatics/farm_rothamsted/issues/634)
+- Require plot number and plot ID for geometry uploads. [#632](https://github.com/Rothamsted-Ecoinformatics/farm_rothamsted/issues/632)
 
 ## [2.19.0](https://github.com/Rothamsted-Ecoinformatics/farm_rothamsted/milestone/40) 2024-02-23